Remove type-specific known failure responses
