+++ tests/Unit/Resque/SendResponseJobTest.php
@@ -2,10 +2,8 @@
 
 namespace App\Tests\Unit\Resque;
 
-use App\Entity\CachedResource;
 use App\Model\Response\AbstractResponse;
-use App\Model\Response\ConnectionFailureResponse;
-use App\Model\Response\HttpFailureResponse;
+use App\Model\Response\KnownFailureResponse;
 use App\Model\Response\SuccessResponse;
 use App\Model\Response\UnknownFailureResponse;
 use App\Resque\Job\SendResponseJob;
@@ -17,7 +15,7 @@ class SendResponseJobTest extends \PHPUnit\Framework\TestCase
      *
      * @param AbstractResponse $response
      */
-    public function testCreateFoo(AbstractResponse $response)
+    public function testCreate(AbstractResponse $response)
     {
         $responseJson = json_encode($response);
 
@@ -55,13 +53,13 @@ class SendResponseJobTest extends \PHPUnit\Framework\TestCase
                 'response' => new UnknownFailureResponse('request_hash_1'),
             ],
             'connection failure response' => [
-                'response' => new ConnectionFailureResponse('request_hash_2', 28),
+                'response' => new KnownFailureResponse('request_hash_2', KnownFailureResponse::TYPE_CONNECTION, 28),
             ],
             'http failure response' => [
-                'response' => new HttpFailureResponse('request_hash_3', 404),
+                'response' => new KnownFailureResponse('request_hash_3', KnownFailureResponse::TYPE_HTTP, 404),
             ],
             'success response' => [
-                'response' => new SuccessResponse('request_hash_4', new CachedResource()),
+                'response' => new SuccessResponse('request_hash_4'),
             ],
         ];
     }
